requirements.txt
pyyaml>=6.0
jsonschema>=4.17.0
pytest>=7.0.0
pytest-cov>=4.0.0

